Fold `BatchIterable` into `ExperimentDataset`
