+++ src/tiledbsoma_ml/_csr.py
@@ -12,7 +12,7 @@ import numpy.typing as npt
 from scipy import sparse
 from typing_extensions import Self
 
-from tiledbsoma_ml._utils import NDArrayNumber
+from tiledbsoma_ml.common import NDArrayNumber
 
 _CSRIdxArray = npt.NDArray[np.unsignedinteger[Any]]
